+++ lib/modules/backupAndRestore/providers/File.js
@@ -1,8 +1,8 @@
 /* @flow */
 
-import { click } from '../../utils';
+import { click } from '../../../utils';
+import { serialize, deserialize } from '../serialization';
 import Provider from './Provider';
-import { serialize, deserialize } from './serialization';
 
 export default class File extends Provider {
 	static text = 'file';
