+++ lib/modules/backupAndRestore/providers/GoogleDrive.js
@@ -1,7 +1,7 @@
 /* @flow */
 
-import { ajax, launchAuthFlow, Permissions } from '../../environment';
-import { string } from '../../utils';
+import { ajax, launchAuthFlow, Permissions } from '../../../environment';
+import { string } from '../../../utils';
 import Provider from './Provider';
 
 const FILE = 'res-storage.json';
