+++ lib/modules/backupAndRestore/providers/Provider.js
@@ -3,7 +3,7 @@
 /* eslint-disable require-await */
 /* eslint-disable no-empty-function */
 
-import * as Metadata from '../../core/metadata';
+import * as Metadata from '../../../core/metadata';
 
 export default class Provider {
 	static text = 'Abstract';
